fix(hir): adapt to TSESTree-tightened EsTreeNode (post-#235)
PR #235 dropped the [key: string]: any escape hatch from EsTreeNode,
so the HIR lower pass — which walks ESTree as a generic graph and
reads properties dynamically — no longer type-checks against the
narrowed Node type.

Two changes:
1. lower.ts: introduce a local `AnyNode = EsTreeNode & Record<string,
   any>` alias and use it throughout the lowering passes. Runtime
   `node.type === "X"` checks still gate every branch — only the
   compiler-time access is widened.
2. runner.ts: tighten the visitor signatures to EsTreeNodeOfType<"...">
   so the per-visitor calls compile against the narrowed shape.

All 744 tests pass; lint, typecheck, format clean.
