+++ packages/oxlint-plugin-react-doctor/src/plugin/hir/index.ts
@@ -1,4 +1,4 @@
 export * from "./types.js";
 export { lowerFunction } from "./lower.js";
 export { inferTypes } from "./infer-types.js";
-export { hirNoSetStateInEffect, hirNoDerivedComputationsInEffects } from "./runner.js";
+export { getOrLowerHir, resolveReportNode } from "./runner.js";
