+++ packages/oxlint-plugin-react-doctor/src/plugin/hir/infer-types.ts
