+++ packages/oxlint-plugin-react-doctor/src/plugin/hir/validators/validate-no-derived-computations-in-effects.ts
