+++ packages/oxlint-plugin-react-doctor/src/plugin/hir/validators/validate-no-set-state-in-effect.ts
